# Optimize Color equality checks