## What does the pull request do?
<!--- Give a bit of background on the PR here, together with links to with related issues etc. -->
Implements `IEquatable` for `Color` struct to optimize equality checks.

## What is the current behavior?
<!--- If the PR is a fix, describe the current incorrect behavior, otherwise delete this section. -->
2 extra allocations per equality check plus quite some overhead.

| Method |     Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------- |---------:|----------:|----------:|-------:|------:|------:|----------:|
|   Test | 39.86 ns | 0.0723 ns | 0.0603 ns | 0.0610 |     - |     - |      48 B |

## What is the updated/expected behavior with this PR?
<!--- Describe how to test the PR. -->
| Method |     Mean |     Error |    StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------- |---------:|----------:|----------:|------:|------:|------:|----------:|
|   Test | 13.13 ns | 0.2594 ns | 0.3883 ns |     - |     - |     - |         - |